requirement.txt
streamlit
langchain-community
langchain_core
python-dotenv
langchain
langchain-groq
fastapi
uvicorn
sse_starlette
SpeechRecognition
pyaudio
gtts
pygame
faiss-cpu
mysql-connector-python
translate
langchain_groq
chromadb